Update userName to be case insensitive with UPPER
This commit updates the SQL transpiler to produce SQL that matches on
userName in a case insensitive manner. This brings the behavior into
compliance with the RFC:
https://datatracker.ietf.org/doc/html/rfc7643#section-4.1.1

For example, prior to this commit a query like the following would not
find users by the username "Bjenson". After this commit, such users
would be returned.

    userName eq "bjenson"

The comparisons are now done by first passing the field and the value
through the SQL function `UPPER`.

These changes only effect the SQL transpiler as the Django Q transpiler
is already case insensitive.

BREAKING CHANGE: This allows queries that did not match rows before to
match rows now!