Migrate to Poetry
This commit updates the package to use Poetry and its related tooling
for development. Changes include, but are not limited to, the addition
of a pyproject.toml file, pytest configurations, the removal of the
setup.py script, and updates to the tox.ini file.

Support for Python 3.6 was is dropped in this commit.