Autoformat Python source with ruff format
Almost all of this repository was *already* using Black-style
formatting, despite there not actually being enforcement of style in CI.

Ruff's autoformatter is outrageously fast and (with small differences)
produces code formatted in the same style as Black.

Let's turn on the autoformatter!

This change was produced by:

1. Editing `pyproject.toml` and `tox.ini` to opt into `ruff format`
2. Changing `noqa F821` to `noqa: F821` (missing the colon means that
   *all* lint rules are ignored)
3. Running `ruff format .`
4. Moving `noqa` comments as necessary!
   ```
   ruff --fix .
   ruff --add-noqa .
   ```

Unfortunately, ruff doesn't allow telling the autoformatter to skip an
enforcement of quote style, so this diff is largely just changing quote
characters.